move OAuthProviderController
